+++ src/Api/Providers/Boards.php
@@ -39,7 +39,9 @@ class Boards extends Provider
      */
     public function forUser($username)
     {
-        return $this->execGetRequest(['username' => $username], UrlHelper::RESOURCE_GET_BOARDS);
+        return $this->execGetRequest(
+            ['username' => $username], UrlHelper::RESOURCE_GET_BOARDS
+        );
     }
 
     /**
@@ -58,9 +60,7 @@ class Boards extends Provider
             'field_set_key' => 'detailed',
         ];
 
-        return $this->execGetRequest(
-            $requestOptions, UrlHelper::RESOURCE_GET_BOARDS
-        );
+        return $this->execGetRequest($requestOptions, UrlHelper::RESOURCE_GET_BOARDS);
     }
 
     /**